Fix some magic numbers and long methods in profile and lesson modules
