+++ features/profile/src/main/kotlin/com/teacherworkout/features/profile/settings/DeleteAccountDialog.kt
@@ -7,23 +7,13 @@ import androidx.compose.ui.res.stringResource
 import com.teacherworkout.features.profile.R
 
 @Composable
-fun DeleteAccountDialog(
-    isOpen: Boolean,
-    onCloseRequest: () -> Unit,
-    onDismissRequest: () -> Unit,
-    onDelete: () -> Unit
-) {
+fun DeleteAccountDialog(onDialogResult: (DialogResult) -> Unit) {
     GenericDialog(
-        isOpen = isOpen,
-        onCloseRequest = onCloseRequest,
-        onDismissRequest = onDismissRequest,
         iconImageVector = Icons.Outlined.Warning,
         iconContentDescription = stringResource(R.string.question_delete_description),
-        mainText = stringResource(R.string.question_delete_account),
-        secondaryText = stringResource(R.string.warning_delete_account),
-        actionText = stringResource(R.string.label_delete_account)
-    ) {
-        onCloseRequest()
-        onDelete()
-    }
+        title = stringResource(R.string.question_delete_account),
+        bodyText = stringResource(R.string.warning_delete_account),
+        actionText = stringResource(R.string.label_delete_account),
+        onDialogResult = onDialogResult
+    )
 }
